clipkit-app/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
python-multipart
python-dotenv
alembic
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator 
httpx
psycopg2-binary
stripe
redis
celery
sentry-sdk[fastapi]
